Revert "feat: stockfish integration"
