+++ src/hooks/useChess.tsx
@@ -1,86 +1,10 @@
 import { useState } from 'react';
-import React from 'react'
 import { Chess, ChessInstance, Move, ShortMove, Square } from 'chess.js';
 import { evaluateBoard } from '@/helpers/chess';
-import { moveMessagePortToContext } from 'worker_threads';
 
 export type ChessType = 'random' | 'computer' | 'minimax';
 
-enum ConnectionStatus {
-  Open = "open",
-  Connecting = "connecting",
-  Disconnected = "disconnected",
-}
-
 const useChess = (type: ChessType) => {
-  // Start of socket
-  // const [socket] = React.useState(io(${process.env.NEXT_PUBLIC_SOCKET_URL}));
-  const [socket, setSocket] = React.useState<WebSocket | null>(null);
-  const [messages, setMessages] = React.useState<string[]>([]);
-  const [connectionStatus, setConnectionStatus] =
-    React.useState<ConnectionStatus>(ConnectionStatus.Disconnected);
-
-  const disconnectSocket = React.useCallback(() => {
-    socket?.close();
-    setSocket(null);
-    setConnectionStatus(ConnectionStatus.Disconnected);
-  }, [socket, setConnectionStatus]);
-
-  React.useEffect(() => {
-    if (socket) {
-      socket.onopen = () => {
-        setConnectionStatus(ConnectionStatus.Open);
-      };
-
-      socket.onmessage = (msg) => {
-        const message = JSON.parse(msg.data);
-        // console.log("Message from BE: ", message);
-        console.log("Message from BE: ", message['message']);
-        // makeMove(message['message'])
-        // console.log(move)
-        // const move = makeMove(message)
-        // const move = makeMove({
-        //   from: '',
-        //   to: targetSquare,
-        //   promotion: 'q',
-        // });
-        makeMove({
-          from: message['message']['from'],
-          to: message['message']['to'],
-          promotion: 'q',
-        });
-        // makeMove(move)
-        setMessages((prev) => [...prev, message.message ?? "EMPTY"]);
-      };
-
-      socket.onerror = () => {
-        disconnectSocket();
-        console.error("Websocket connection error");
-      };
-
-      socket.onclose = () => {
-        disconnectSocket();
-      };
-    }
-  }, [disconnectSocket, setConnectionStatus, socket]);
-
-  const sendMessage = (move: string) => {
-    console.log("MOVE: ", move);
-    socket?.send(move);
-  };
-
-  const onCloseSocket = () => {
-    socket?.close();
-    setSocket(null);
-    setConnectionStatus(ConnectionStatus.Disconnected);
-  };
-
-  const connectSocket = () => {
-    const url = `${process.env.NEXT_PUBLIC_SOCKET_URL}` + "/ws";
-    const socket = new WebSocket(url);
-    setSocket(socket);
-  };
-  // End of socket
   const [game, setGame] = useState(new Chess());
   const [playing, setPlaying] = useState(false);
   const [moves, setMoves] = useState<Move[]>([]);
@@ -96,7 +20,6 @@ const useChess = (type: ChessType) => {
   };
 
   const makeMove = (move: string | ShortMove) => {
-    console.log(move)
     const gameCopy = { ...game };
     const result = gameCopy.move(move);
 
@@ -135,7 +58,6 @@ const useChess = (type: ChessType) => {
     }
 
     if (bestMove === null) return;
-    console.log(bestMove)
     makeMove(bestMove);
   };
 
@@ -214,13 +136,9 @@ const useChess = (type: ChessType) => {
     });
 
     if (move === null) return false;
-    sendMessage(move.from + move.to)
-    console.log(move)
-    console.log("My move: ")
-    console.log(move.from + move.to)
 
-    // const newTimeout = setTimeout(getComputerType(), 500);
-    // setCurrentTimeout(newTimeout);
+    const newTimeout = setTimeout(getComputerType(), 500);
+    setCurrentTimeout(newTimeout);
     return true;
   };
 
@@ -236,7 +154,6 @@ const useChess = (type: ChessType) => {
   };
 
   const startGame = () => {
-    connectSocket()
     resetGame();
     setPlaying(true);
   };
